Factor out even more functionality, reducing the size of the binaries and tests, and parallelizing both small and medium tests.
